+++ src/models/InlineStyle.js
@@ -34,7 +34,16 @@ export default class InlineStyle {
           console.warn(`Node of type ${node.type} not supported as an inline style`)
         }
       })
-      const styleObject = transformDeclPairs(declPairs)
+      // RN currently does not support differing values for the corner radii of Image
+      // components (but does for View). It is almost impossible to tell whether we'll have
+      // support, so we'll just disable multiple values here.
+      // https://github.com/styled-components/css-to-react-native/issues/11
+      const styleObject = transformDeclPairs(declPairs, [
+        'borderRadius',
+        'borderWidth',
+        'borderColor',
+        'borderStyle',
+      ])
       const styles = StyleSheet.create({
         generated: styleObject,
       })